大事记：修复按年份筛选结果全空（年份列浮点化）
date 列有任意一条解析失败时，.dt.year 退化成浮点，astype(str)
得到 "2025.0" 与下拉框的 "2025" 永不相等。改用 Int64 再转字符串。

+++ app_pages/1_📅_大事记.py
@@ -254,7 +254,8 @@ except Exception as e:
 # 年份筛选
 _all = (t("all"),)
 if not df.empty and year_filter not in ("全部", "All"):
-    df["year"] = pd.to_datetime(df["date"], errors="coerce").dt.year.astype(str)
+    # 用 Int64 可空整数再转字符串，避免有 NaT 时整列退化成浮点（"2025.0" ≠ "2025" 导致筛空）
+    df["year"] = pd.to_datetime(df["date"], errors="coerce").dt.year.astype("Int64").astype(str)
     df = df[df["year"] == year_filter]
 
 if not df.empty and tag_filter not in ("全部", "All"):
